implemented input and logging system + various other tweaks
- included RapidJSON 1.1.0
- implemented input system
- updated application quit keybind from escape to ctrl+q
- added timer
- added playerprefs
- added flexlogger
- created file handler for flexlogger
- expanded shaders into a cpp file
- added flow logging
- updated pch
- removed abort() in place of Log::Fatal to attempt to allow the application to free memory (if it crashes anyways, too bad)
- FATAL ERROR HANDLING IS UNTESTED

+++ FlexEditor/src/Editor/layer_editor.cpp
@@ -226,7 +226,7 @@ namespace FlexEngine
     ImGui::Begin("Viewport");
     ImGui::Text("pointer = %p", test_image.texture);
     ImGui::Text("size = %d x %d", test_image.width, test_image.height);
-    ImGui::Image((void*)(intptr_t)test_image.texture, ImVec2(test_image.width, test_image.height));
+    ImGui::Image((void*)(intptr_t)test_image.texture, ImVec2(static_cast<float>(test_image.width), static_cast<float>(test_image.height)));
     ImGui::End();
 
     ImGui::ShowDemoWindow();
